+++ src/TestUtilOutputGenerator_Consumer/TestUtilOutputGeneratorConsumerProgram.cs
@@ -16,10 +16,9 @@ namespace TestUtilOutputGenerator_Consumer
             TestUtilRunner runner = new TestUtilRunner();
 
             string basepath = "";
-            //basepath = DebugOnlyHelper.DebugOnlyScriptFolderHelper.TestingScriptDirectory();
 
-            //basepath = @"C:\dev\gitrepos\testutil\src\ScriptsForTesting\Test1";
-            basepath = @"C:\dev\gitrepos\testutil\scripts";
+            basepath = @"C:\dev\gitrepos\testutil\src\ScriptsForTesting\Test1";
+            //basepath = @"C:\dev\gitrepos\testutil\scripts";
 
 
             Task<Report> task = RunEngineAsync(runner, basepath);